Implement interactive list
